extract seed 36 component
