+++ src/components/seed/36/index.ts
@@ -1,0 +1,1 @@
+export { default as Seed36Steps } from './Seed36Steps';
